Safari-611.1.10/LayoutTests/fast/images/cmyk-jpeg-with-color-profile-expected.txt

Test that we can render a CMYK JPEG without color corruption.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS green > red is true
PASS green > blue is true
PASS successfullyParsed is true

TEST COMPLETE
